platform/qemu: move disk-related code into disk.go
